Changed file names and added 2 classes
