+++ pyaptamer/aptanet/tests/test_aptanet.py
@@ -4,7 +4,8 @@ import numpy as np
 import pytest
 from sklearn.utils.estimator_checks import parametrize_with_checks
 
-from pyaptamer.aptanet import AptaPipeline
+from pyaptamer.aptanet import AptaNetPipeline
+from pyaptamer.aptanet._pipeline import AptaNetFeaturesClassifier
 
 
 @pytest.mark.skipif(
@@ -23,7 +24,7 @@ def test_pipeline_fit_and_predict(aptamer_seq, protein_seq):
     """
     Test if Pipeline predictions are valid class labels and shape matches input.
     """
-    pipe = AptaPipeline()
+    pipe = AptaNetPipeline()
 
     X_raw = [(aptamer_seq, protein_seq) for _ in range(40)]
     y = np.array([0] * 20 + [1] * 20, dtype=np.float32)
@@ -38,7 +39,7 @@ def test_pipeline_fit_and_predict(aptamer_seq, protein_seq):
 @pytest.mark.skipif(
     sys.version_info >= (3, 13), reason="skorch does not support Python 3.13"
 )
-@parametrize_with_checks([AptaPipeline()])
+@parametrize_with_checks([AptaNetFeaturesClassifier()])
 def test_sklearn_compatible_estimator(estimator, check):
     """
     Run scikit-learn's compatibility checks on the AptaPipeline.
